// src/routes/[tenant]/dashboard/supplies/categories/schema.ts
import type { Description } from '$lib/components/ui/sheet';
import { z } from 'zod/v4';

export const add = z.object({
	name: z.string('Name of Payment Method is required').min(2).max(50),
	description: z.string('Description is required').min(2).max(100)
});

export const edit = z.object({
	id: z.coerce.string(),
	name: z.string('Name of Payment Method is required').min(2).max(50),
	description: z.string('Description is required').min(2).max(100)
});
export type Edit = z.infer<typeof edit>;